refactor: split transport probe behind facade
NavigationTransportProbe.ts had grown to 1036 LoC, the largest file
in the Mediator/Init cluster. Split it into focused co-located
modules under TransportProbe/ (Types, Reject, Result, Url, Dns, Tcp,
Tls, Probe) and turn the original file into a thin barrel facade that
re-exports the unchanged public surface.

Pure behaviour-preserving code-motion: the probe unit test passes
unmodified (18 cases) and NavigationDiagnostics keeps importing the
same names. Prepares the Init cluster for the per-file size cap.

Docs: the init-navigation-forensics page notes the facade split; the
24 cross-module-internal exports forced by the split are allowlisted
for docs-coverage (the public probe surface stays documented).
